+++ src/main/java/com/ricedotwho/rsm/ui/clickgui/impl/module/settings/ValueComponent.java
@@ -30,9 +30,9 @@ public abstract class ValueComponent<T extends Setting<?>> {
 
     public abstract void render(GuiGraphics gfx, double mouseX, double mouseY, float partialTicks);
 
-    public abstract void click(double mouseX, double mouseY, float mouseButton);
+    public abstract void click(double mouseX, double mouseY, int mouseButton);
 
-    public abstract void release(double mouseX, double mouseY, float mouseButton);
+    public abstract void release(double mouseX, double mouseY, int mouseButton);
 
     public boolean key(char typedChar, int keyCode) {
         return false;
