+++ src/main/java/com/ricedotwho/rsm/ui/clickgui/impl/module/settings/impl/EmptyValueComponent.java
@@ -16,12 +16,12 @@ public class EmptyValueComponent extends ValueComponent<Setting<?>> {
     }
 
     @Override
-    public void click(double mouseX, double mouseY, float mouseButton) {
+    public void click(double mouseX, double mouseY, int mouseButton) {
         // intentionally empty
     }
 
     @Override
-    public void release(double mouseX, double mouseY, float mouseButton) {
+    public void release(double mouseX, double mouseY, int mouseButton) {
         // intentionally empty
     }
 }
